# Core Audio for interfaces with large #'s of channels