@dholl 2017-04-17
For the two API changes in my current commits, here are two alternative implementations that cross my mind:

1)  ```get_channel_name```
  Would this functionality be better merged into with query_devices?  For example, in the dict returned for each device, perhaps each dict could contain additional fields called "input_channel_names" and "output_channel_names" which each contain a list (or tuple) of strings, 1 string per channel.  This way, hostapi's such as ASIO and Core Audio could be queried for whatever underlying names.  (For ASIO, is the equivalent underlying API "PaAsio_GetOutputChannelName" and "PaAsio_GetInputChannelName"?)

2)  ```MacCoreSettings```
    This struct has at least field in common with ```AsioSettings```, specifically ```channel_selectors``` which is the only attribute presently exposed in ```AsioSettings```.  Regarding the Core Audio API, I considered calling this field ```channel_map``` in ```MacCoreSettings```, but upon reading how similar in functionality it is to the field in ```AsioSettings```, I added some minor conversion logic to translate ```channel_selectors``` into the specific format that Core Audio required -- as mentioned in my code's doc string.
    Could all 3 of the API-specific settings classes (```AsioSettings, MacCoreSettings, WasapiSettings```) be merged into 1 simple dict?  Presently, a client application needs to select the correct ```...Settings``` class based on the relevant device's ```hostapi```, correct?  Instead of this, how about we get rid of the classes in favor of having the application just pass in a simple dict instead?
    ```
    extra_settings_i = {
    'channel_selectors' = [8, 9], # RME UFX+ 2 mic inputs
    'change_dev_params' = True, # Let Core Audio (or any underlying API) change hardware params as needed to minimize host overhead
    'allow_rate_conv' = False, # I would rather fail and alert the user than let underlying layer (Core Audio) have to resort to sample rate conversion.
    }
    extra_settings_o = {
    'channel_selectors' = [0, 1], # RME UFX+ first 2 outputs
    'change_dev_params' = True, # same as extra_settings_i
    'allow_rate_conv' = False,  # same as extra_settings_i
    }
    with sd.RawStream(device=..., extra_settings=(extra_settings_i, extra_settings_o), ...)
      ...
    ```
    These dicts could then be passed to a correct hostapi-specific class or function from within ```_get_stream_parameters```.  For example:
    ```
    def _get_stream_parameters(kind, device, channels, dtype, latency,
                               extra_settings, samplerate):
      ...
      if samplerate is None:
          samplerate = info['default_samplerate']
      if extra_settings:
          # TODO: build dict extra_settings_xlate to map hostapi to a callable to translate extra_settings into hostapi-specific _streaminfo.
          #extra_settings_xlate = {
          #    ...
          #}
          # Invoke the callable (function or object) that is specific for this hostapi:
          this_streaminfo = extra_settings_xlate[info['hostapi']](kind=kind, device=device, channels=channels, sampleformat=sampleformat, latency=latency, extra_settings=extra_settings)
      else:
          this_streaminfo = _ffi.NULL
      parameters = _ffi.new('PaStreamParameters*', (
          device, channels, sampleformat, latency,
          this_streaminfo))
    ```
    * Aside: Doing so would also make the implementation of ```MacCoreSettings``` a little clearer, since I require access to the device's ```max_output_channels``` to translate ```channel_selectors``` into Core Audio's ```channel_map```.  At the moment, I defer creating a ```PaMacCoreStreamInfo``` struct until ```_get_stream_parameters```, be having that function discover that ```MacCoreSettings._streaminfo``` is callable and then invoking it with device-specific params passed in via **kwargs.
        * I'll admit this is implementation is more complicated that I care for, but at least if an application tries using my current revision of ```MacCoreSettings```, then it doesn't have to explicitly care about ```max_output_channels```, and it may use use the same format of ```channel_selectors``` for both ```MacCoreSettings``` and ```AsioSettings```.
    * So to keep things simple in my initial pull request, I just created ```MacCoreSettings``` in a similar fashion to ```AsioSettings``` and ```WasapiSettings```, to keep my changes as similar to the rest of python-sounddevice as possible.

If either or both of these two ideas don't sound too ugly, I'm happy to conjure a first draft so that you'd have something concrete to ponder.