@tgarc 2017-04-17
Most of all I'd like to get [GetAvailableBufferSizes](http://portaudio.com/docs/v19-doxydocs/pa__asio_8h.html#a95dbac9747083dcee51b1e0e4695a912) added. But GetInputChannelName/GetOutputChannelName could also be useful.