@dholl 2017-04-17
Hmm, and Core Audio offers similar data via kAudioDevicePropertyBufferFrameSizeRange, which is exposed via [```PaMacCore_GetBufferSizeRange```](http://portaudio.com/docs/v19-doxydocs/pa__mac__core_8h.html#aa470cfe0a7a53a6bc7eab4b18f670d31), though this call does not return ```preferredBufferSizeFrames``` or ```granularity```.

Perhaps could this data also be returned by ```query_devices``` along with channel names?  Maybe via a dict key such as:
```
{
'buffer_sizes': range(minBufferSizeFrames, maxBufferSizeFrames, granularity) if granularity!=-1 else power_of_two_range(minBufferSizeFrames, maxBufferSizeFrames),
}
#TODO: define power_of_two_range
```
where the dict value contains some "iterable" and doesn't have to be a list or tuple.  (An app is free to fee this iterable to ```list()``` or ```tuple()``` or whatever as desired.)  To prevent accidental expansion of the range of values, ```xrange``` would be used on Python 2, whereas Python 3 can stick with just ```range```.

On Core Audio, I *think* ```granularity``` may be hard coded to +1 (is there a way to tell this?), but I'm not sure what's the equivalent to ```preferredBufferSizeFrames```.  Is ```preferredBufferSizeFrames``` desirable, or could knowledge of ```preferredBufferSizeFrames ``` be incorporated into the library in another fashion such as setting the default latency or equivalent?