@tgarc 2017-04-18
I wouldn't get too attached to the idea of exposing host api buffer sizes as some kind of common functionality. Portaudio is pretty generic at the stream level so the feature set across apis is pretty damn varied. For one thing, ALSA doesn't have any feature like this - which is the only real api that it is supported in Linux. Another example is PaWasapi's [GetFramesPerHostBuffer](http://portaudio.com/docs/v19-doxydocs/pa__win__wasapi_8h.html#af8f04003c6123e6fdcfdf7308761a75e) which is for a similiar use case as GetAvailableBufferSizes but is more limited. I agree It would be nice if the set and get interface for similar features were functionally identical (same input variables, same output variables) but I'm not sure how much value it would provide given the lack of similarity between apis.

One other thing to keep in mind in this discussion: not all host specific settings are stream specific. For example Alsa has a few [settings](http://portaudio.com/docs/v19-doxydocs/pa__linux__alsa_8h.html#a76f60d6f1ba6a298c038a9f3a6dccc53) that apply to all created streams.