@dholl 2017-04-24
@mgeier
> But there is still one question open: does this stuff actually work?

I *hope* so.  ;)

> I tried the channel_map a bit, and the output mapping seems to work. I assume the input works, too, but I didn't fully check this (it gave me one channel, but I didn't check if it was the one I asked for).

I tested with 2 mic's plugged into different ports on a UFX+, and it does let me switch.

> How can I check if change_device_parameters does anything?

When I use it, it changes ```default_samplerate``` as returned from ```query_devices()```.  Here's a device list with my laptop's built-in hardware and my own test application:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                    
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      48000.0       |         2          |    0.012104166666666666    |   0.0027708333333333335   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      48000.0       |         0          |            0.1             |           0.01            |          2          |    0.020020833333333335     |         0.0106875

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |
          1 | ''    |

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''
          1 |       | ''
```

Now, run at 96 kHz without ```change_device_parameters```.  Here I'm inputting 1 channel from the microphone, and outputting to the right speaker:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 96000
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.10139583333333334, 0.03536458333333333)
time = 342188.05329327105
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 96000.0
samplesize = (4, 4)
```

And list devices again:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                                    
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      48000.0       |         2          |    0.012104166666666666    |   0.0027708333333333335   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      48000.0       |         0          |            0.1             |           0.01            |          2          |    0.020020833333333335     |         0.0106875         

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |       
          1 | ''    |       

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''    
          1 |       | ''    
```

Now we'll do the same run again, but this time, my ```--tweakdev 1``` enables ```change_device_parameters```:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 --tweakdev 1 -s 96000
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.04921875, 0.019135416666666665)
time = 342375.305821312
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 96000.0
samplesize = (4, 4)
```

And when we re-run ```query_devices()```, we see that ```default_samplerate``` is now 96 kHz:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                                                 
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      96000.0       |         2          |    0.006552083333333333    |   0.0018854166666666668   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      96000.0       |         0          |            0.1             |           0.01            |          2          |    0.013802083333333333     |    0.009135416666666667   

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |       
          1 | ''    |       

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''    
          1 |       | ''    
```

I also notice subtle pops from the speakers when I'm changing the sampling rate with ```change_device_parameters```, but if I specify the same sampling rate, using ```change_device_parameters``` does not introduce any pops on my laptop.   (late 2016 MacBook Pro)

> How can I provoke a failure with fail_if_conversion_required?

Trying 8 kHz without ```change_device_parameters``` works fine, with a subtle change in sound when just rubbing my finger over the mic location on my laptop:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000              
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.05025, 0.108875)
time = 342676.486970847
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)
```

Note the default sample rate is still 96000:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                                   
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      96000.0       |         2          |    0.006552083333333333    |   0.0018854166666666668   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      96000.0       |         0          |            0.1             |           0.01            |          2          |    0.013802083333333333     |    0.009135416666666667   

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |       
          1 | ''    |       

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''    
          1 |       | ''    
```
(Note: Core Audio returns empty strings for channel names of builtin inputs and outputs.  But on a UFX+, I do get useful strings such as 'Mic 9' and whatnot.)

So now. with ```change_device_parameters```, the same 8 kHz request produces a couple error messages, but playback continues as expected:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 --tweakdev 1 -s 8000
||PaMacCore (AUHAL)|| Error on line 406: err=''!dat'', msg=Audio Device: Unsupported Format
||PaMacCore (AUHAL)|| Error on line 406: err=''!dat'', msg=Audio Device: Unsupported Format
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 342830.96096854704
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)
```

But now the default_sample_rate is 44.1 kHz:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                                                
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      44100.0       |         2          |    0.01310657596371882     |   0.0029478458049886623   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      44100.0       |         0          |            0.1             |           0.01            |          2          |    0.021156462585034015     |    0.010997732426303855   

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |       
          1 | ''    |       

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''    
          1 |       | ''    
```

So running one last time to test ```fail_if_conversion_required```, my equivalent command line option is ```--norateconv 1```:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 --tweakdev 1 --norateconv 1 -s 8000
||PaMacCore (AUHAL)|| Error on line 406: err=''!dat'', msg=Audio Device: Unsupported Format
Traceback (most recent call last):
  File "./dholl_wire_sounddevice.py", line 302, in <module>
    _main()
  File "./dholl_wire_sounddevice.py", line 280, in _main
    channels=args.channels, callback=callback, clip_off=not args.clip, dither_off=not args.dither) as s:
  File "/Users/dholl/Library/Python/2.7/lib/python/site-packages/sounddevice.py", line 1551, in __init__
    **_remove_self(locals()))
  File "/Users/dholl/Library/Python/2.7/lib/python/site-packages/sounddevice.py", line 1077, in __init__
    'Error opening {0}'.format(self.__class__.__name__))
  File "/Users/dholl/Library/Python/2.7/lib/python/site-packages/sounddevice.py", line 3149, in _check
    raise PortAudioError(msg)
sounddevice.PortAudioError: Error opening RawStream: Invalid sample rate
zsh: exit 1     ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1  1 
```
So that's good that we see an error.

However, running  again without ```--tweakdev 1``` but still with ```--norateconv 1```, we find that no error messages are printed nor is any exception raised:
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 --norateconv 1 -s 8000 
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343026.211521728
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)
```

This unintuitive enforcement of ```fail_if_conversion_required``` only when using ```change_device_parameters``` is consistent with the doc string in ```CoreAudioSettings```.  Specifically, the part stating ***In combination with the above flag, ...***:
```
        change_device_parameters : bool, optional
            If ``True``, allows PortAudio to change things like the
            device's frame size, which allows for much lower latency,
            but might disrupt the device if other programs are using it,
            even when you are just querying the device.  ``False`` is
            the default.
        fail_if_conversion_required : bool, optional
            In combination with the above flag, ``True`` causes the
            stream opening to fail, unless the exact sample rates are
            supported by the device.
```

So from a user application standpoint, if it wants ```fail_if_conversion_required=True```, then it must also set ```change_device_parameters```.

> I think I'm too lazy to investigate if conversion_quality works, did you check this?
Ha!  Me to...

I wasn't in the mood today for running any spectral analysis such as identifying the phase/magnitude response or intermodulation distortion.  Though, I could conceivable do something with getting a UFX+ to loop back my mac's output to software.

For now, without having a paying contract in place  ;)  I've only run my tool with each of the 5 conversion algorithms, and they each sound too similar to my untrained ears at conversion from 44.1 kHz to 8 kHz.  So, all this verified is that the underlying Port Audio and Core Audio layers are not vomiting.
```
zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000 --convquality min
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343735.471557744
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)

zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000 --convquality low
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343741.264996927
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)

zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000 --convquality medium
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343972.305505788
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)

zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000 --convquality high
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343747.151818073
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)

zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -c 1 --input-chan-map _0 --output-chan-map _1 -s 8000 --convquality max   
################################################################################
press Return to quit
################################################################################
cpu_load = 0.0
latency = (0.077, 0.062125)
time = 343976.846742895
blocksize = 0
channels = (1, 1)
device = (0, 1)
dtype = ('float32', 'float32')
samplerate = 8000.0
samplesize = (4, 4)

zero(ttys002):...Python/sound> ./dholl_wire_sounddevice.py -L                                                                     
Device list:
ID |         name          |   hostapi    | default_samplerate | max_input_channels | default_high_input_latency | default_low_input_latency | max_output_channels | default_high_output_latency | default_low_output_latency
---+-----------------------+--------------+--------------------+--------------------+----------------------------+---------------------------+---------------------+-----------------------------+---------------------------
 0 | 'Built-in Microphone' | 'Core Audio' |      44100.0       |         2          |    0.01310657596371882     |   0.0029478458049886623   |          0          |             0.1             |            0.01           
 1 |   'Built-in Output'   | 'Core Audio' |      44100.0       |         0          |            0.1             |           0.01            |          2          |    0.021156462585034015     |    0.010997732426303855   

Channel names for device 0, Built-in Microphone:
        Ch# | Input | Output
        ----+-------+-------
          0 | ''    |       
          1 | ''    |       

Channel names for device 1, Built-in Output:
        Ch# | Input | Output
        ----+-------+-------
          0 |       | ''    
          1 |       | ''    
```
Hmm, looking at my CPU load of coreaudiod, the load doesn't seem to vary much between max, high, medium, low, or min.  All are around 25-27%